// src/data/cards.generated.ts
export interface RawCard {
  id: string;
  name?: string;
  attribute?: string;
  tooltip?: string;
  phys_atk?: number | null;
  mind_atk?: number | null;
  mp_cost?: number | null;
  type?: string;
  tags?: string | null;
  chainable?: boolean | null;
  chain_value?: number | null;
  target?: string | null;
  [k: string]: any;
}